Set default value for new property vertical accuracy
:Release Notes:
Add default value for vertical accuracy

:Detailed Notes:
We have requirement to data also with getLocationUpdates API response,
which was currently set to default value '-1'. So to support the same
added a new parameter in nyx_gps_location_t data structure and here we
set it with the default value -1 menas unknow value.

:Testing Performed:
Build Success

:QA Notes:
N/A

:Issues Addressed:
[PLAT-117491] Correct vertical accuracy value for getLocationUpdates
              API

Change-Id: Iefc3434b5f2e5ee5f5f0fc5898a0d0816852b591